pop front and pop back
