+++ datafusion/physical-expr/src/functions.rs
@@ -258,12 +258,6 @@ pub fn create_physical_fun(
         BuiltinScalarFunction::ArrayExcept => Arc::new(|args| {
             make_scalar_function_inner(array_expressions::array_except)(args)
         }),
-        BuiltinScalarFunction::ArrayPopFront => Arc::new(|args| {
-            make_scalar_function_inner(array_expressions::array_pop_front)(args)
-        }),
-        BuiltinScalarFunction::ArrayPopBack => Arc::new(|args| {
-            make_scalar_function_inner(array_expressions::array_pop_back)(args)
-        }),
         BuiltinScalarFunction::ArrayPosition => Arc::new(|args| {
             make_scalar_function_inner(array_expressions::array_position)(args)
         }),
@@ -290,9 +284,6 @@ pub fn create_physical_fun(
         }),
         BuiltinScalarFunction::ArrayReverse => Arc::new(|args| {
             make_scalar_function_inner(array_expressions::array_reverse)(args)
-        }),
-        BuiltinScalarFunction::ArraySlice => Arc::new(|args| {
-            make_scalar_function_inner(array_expressions::array_slice)(args)
         }),
         BuiltinScalarFunction::ArrayIntersect => Arc::new(|args| {
             make_scalar_function_inner(array_expressions::array_intersect)(args)
