+++ datafusion/proto/proto/datafusion.proto
@@ -641,7 +641,7 @@ enum ScalarFunction {
   // 97 was ArrayToString
   // 98 was Cardinality
   // 99 was ArrayElement
-  ArraySlice = 100;
+  // 100 was ArraySlice
   Cot = 103;
   // 104 was ArrayHas
   // 105 was ArrayHasAny
@@ -655,7 +655,7 @@ enum ScalarFunction {
   // 113 was IsNan
   Iszero = 114;
   // 115 was ArrayEmpty
-  ArrayPopBack = 116;
+  // 116 was ArrayPopBack
   // 117 was StringToArray
   // 118 was ToTimestampNanos
   ArrayIntersect = 119;
@@ -663,7 +663,7 @@ enum ScalarFunction {
   OverLay = 121;
   /// 122 is Range
   ArrayExcept = 123;
-  ArrayPopFront = 124;
+  // 124 was ArrayPopFront
   Levenshtein = 125;
   SubstrIndex = 126;
   FindInSet = 127;
